fix: [session-persistence-default-alias-uuid-mergeability] - Merge main and restore workers scheduler integration
Integrate current main (workers service extraction, backend scope config)
while keeping UUID default-session stream identity; wire workersScheduler
through FactoryService, combine backendRuntimeCacheScope with resolved
stream identity in dashboard bootstrap, and satisfy dashboard folder limits.
