+++ ui/src/features/dashboard/hooks/useDashboardSnapshot.ts
@@ -1,25 +1,18 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef } from "react";
 
 import type { FactoryEvent } from "../../../api/events";
 import {
   clearTimelineCheckpoint,
   type FactoryTimelineCheckpoint,
   persistTimelineCheckpoint,
-  purgeLegacyTimelineCheckpoints,
   reconnectCursorFromCheckpoint,
   type TimelineCheckpointStreamIdentity,
   useFactoryTimelineStore,
 } from "../../timeline/public";
-import {
-  useDashboardSession,
-  useSetDashboardSessionID,
-} from "../session/dashboard-session-provider";
-import {
-  bootstrapDashboardSessionSyncPreflight,
-  type DashboardSessionRecoveryState,
-} from "../lib/dashboard-session-sync-preflight";
+import { useDashboardSession } from "../session/dashboard-session-provider";
 import { useDashboardStreamStore } from "../state/dashboardStreamStore";
 import { useFactoryEventStream } from "./event-stream/useFactoryEventStream";
+import { useGuardedTimelineCheckpointBootstrap } from "./snapshot/useDashboardSnapshot.bootstrap";
 import { useDashboardSessionLifecycle } from "./useDashboardSessionLifecycle";
 import { useDashboardTimelineMemoryDebug } from "./useDashboardTimelineMemoryDebug";
 import { useDashboardWorldView } from "./useDashboardWorldView";
@@ -31,149 +24,6 @@ export interface UseDashboardSnapshotOptions {
 }
 
 type DashboardPreflightStatus = "loading" | "non-recoverable" | "success";
-
-function useGuardedTimelineCheckpointBootstrap({
-  checkpointHydrationKey,
-  checkpointsDisabled,
-  rawSessionID,
-  refreshToken,
-  restoreCheckpoint,
-  setResolvedStreamIdentity,
-}: {
-  checkpointHydrationKey: string | null;
-  checkpointsDisabled: boolean;
-  rawSessionID: string | null;
-  refreshToken: number;
-  restoreCheckpoint: (checkpoint: FactoryTimelineCheckpoint) => void;
-  setResolvedStreamIdentity: (
-    streamIdentity: TimelineCheckpointStreamIdentity | null,
-  ) => void;
-}) {
-  const setSelectedSessionID = useSetDashboardSessionID();
-  const setStreamState = useDashboardStreamStore((state) => state.setStreamState);
-  const [checkpointHydratedKey, setCheckpointHydratedKey] =
-    useState<string | null>(null);
-  const [persistedCheckpoint, setPersistedCheckpoint] =
-    useState<FactoryTimelineCheckpoint | null>(null);
-  const [preflightReadyKey, setPreflightReadyKey] = useState<string | null>(
-    null,
-  );
-  const [preflightError, setPreflightError] = useState<Error | null>(null);
-  const [preflightRecovery, setPreflightRecovery] =
-    useState<DashboardSessionRecoveryState | null>(null);
-  const [streamIdentity, setStreamIdentity] =
-    useState<TimelineCheckpointStreamIdentity | null>(null);
-
-  const checkpointHydrated = checkpointHydratedKey === checkpointHydrationKey;
-  const preflightReady = preflightReadyKey === checkpointHydrationKey;
-
-  useEffect(() => {
-    let cancelled = false;
-
-    setCheckpointHydratedKey(null);
-    setPersistedCheckpoint(null);
-    setPreflightError(null);
-    setPreflightRecovery(null);
-    setPreflightReadyKey(null);
-    setStreamIdentity(null);
-    setResolvedStreamIdentity(null);
-
-    if (
-      checkpointHydrationKey == null ||
-      !rawSessionID ||
-      typeof window === "undefined" ||
-      checkpointsDisabled
-    ) {
-      setPreflightReadyKey(checkpointHydrationKey);
-      setCheckpointHydratedKey(checkpointHydrationKey);
-      return;
-    }
-
-    void bootstrapDashboardSessionSyncPreflight({
-      indexedDB: window.indexedDB,
-      refreshToken,
-      sessionID: rawSessionID,
-    })
-      .then(async (outcome) => {
-        if (cancelled) {
-          return;
-        }
-        if (outcome.kind === "error") {
-          setPreflightError(outcome.error);
-          setStreamState({
-            message: outcome.error.message,
-            status: "offline",
-          });
-          setCheckpointHydratedKey(checkpointHydrationKey);
-          return;
-        }
-        if (outcome.kind === "recovery") {
-          setPreflightRecovery(outcome.recovery);
-          setCheckpointHydratedKey(checkpointHydrationKey);
-          return;
-        }
-
-        const {
-          checkpoint,
-          reconnectCursor: _reconnectCursor,
-          remappedFactorySessionId,
-          streamIdentity: resolvedStreamIdentity,
-        } = outcome.result;
-        setStreamIdentity(resolvedStreamIdentity);
-        setResolvedStreamIdentity(resolvedStreamIdentity);
-        if (
-          remappedFactorySessionId != null &&
-          remappedFactorySessionId !== rawSessionID
-        ) {
-          setSelectedSessionID(remappedFactorySessionId);
-        }
-        await purgeLegacyTimelineCheckpoints(window.indexedDB);
-        setPreflightReadyKey(checkpointHydrationKey);
-        if (checkpoint) {
-          restoreCheckpoint(checkpoint);
-        }
-        setPersistedCheckpoint(checkpoint);
-        setCheckpointHydratedKey(checkpointHydrationKey);
-      })
-      .catch((preflightError: unknown) => {
-        if (cancelled) {
-          return;
-        }
-        const message =
-          preflightError instanceof Error && preflightError.message.trim() !== ""
-            ? preflightError.message
-            : "The dashboard could not load the selected session.";
-        setPreflightError(new Error(message));
-        setStreamState({
-          message,
-          status: "offline",
-        });
-        setCheckpointHydratedKey(checkpointHydrationKey);
-      });
-
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    checkpointHydrationKey,
-    checkpointsDisabled,
-    rawSessionID,
-    refreshToken,
-    restoreCheckpoint,
-    setResolvedStreamIdentity,
-    setSelectedSessionID,
-    setStreamState,
-  ]);
-
-  return {
-    checkpointHydrated,
-    preflightError,
-    preflightRecovery,
-    persistedCheckpoint,
-    preflightReady,
-    streamIdentity,
-  };
-}
 
 function usePersistedTimelineCheckpoint({
   checkpoint,
@@ -218,6 +68,9 @@ export function useDashboardSnapshot({
   const setResolvedStreamIdentity = useDashboardStreamStore(
     (state) => state.setResolvedStreamIdentity,
   );
+  const setBackendRuntimeCacheScope = useDashboardStreamStore(
+    (state) => state.setBackendRuntimeCacheScope,
+  );
   const { error, isInitialLoading, snapshot, streamState } =
     useDashboardWorldView();
   const { isPaused, rawSessionID } = useDashboardSession();
@@ -254,6 +107,7 @@ export function useDashboardSnapshot({
     rawSessionID,
     refreshToken,
     restoreCheckpoint,
+    setBackendRuntimeCacheScope,
     setResolvedStreamIdentity,
   });
