+++ ui/src/features/dashboard/state/dashboardStreamStore.ts
@@ -5,8 +5,10 @@ import type { StreamDerivedCacheIdentity } from "../../timeline/public";
 import { getDashboardStreamMessages } from "../messages/dashboard-stream";
 
 interface DashboardStreamStoreState {
+  backendRuntimeCacheScope: string | null;
   resetStreamState: (locale?: string | null) => void;
   resolvedStreamIdentity: StreamDerivedCacheIdentity | null;
+  setBackendRuntimeCacheScope: (backendRuntimeCacheScope: string | null) => void;
   setResolvedStreamIdentity: (
     streamIdentity: StreamDerivedCacheIdentity | null,
   ) => void;
@@ -25,13 +27,18 @@ export function createDefaultDashboardStreamState(
 
 export const useDashboardStreamStore = create<DashboardStreamStoreState>(
   (set) => ({
+    backendRuntimeCacheScope: null,
     resetStreamState: (locale) => {
       set({
+        backendRuntimeCacheScope: null,
         resolvedStreamIdentity: null,
         streamState: createDefaultDashboardStreamState(locale),
       });
     },
     resolvedStreamIdentity: null,
+    setBackendRuntimeCacheScope: (backendRuntimeCacheScope) => {
+      set({ backendRuntimeCacheScope });
+    },
     setResolvedStreamIdentity: (streamIdentity) => {
       set({ resolvedStreamIdentity: streamIdentity });
     },
